requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
asyncio==4.0.0
attrs==25.3.0
automata-lib==9.1.2
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
cached_method==0.1.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.3
contourpy==1.2.1
cryptography==46.0.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
django-active-link==0.2.2
django-allauth==0.61.0
django-cloudinary-storage==0.3.0
docopt==0.6.2
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.18.0
fonttools==4.51.0
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.45.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1

kiwisolver==1.4.5
lark==1.3.0
lxml==5.2.1
MarkupSafe==3.0.2
mistune==3.1.4
msgpack==1.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
oauthlib==3.3.1
packaging==24.0
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.3.8
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pymunk==6.11.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python3-openid==3.2.0
PyTMX==3.32
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.3
pyzmq==27.1.0
razorpay==2.0.0
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
Send2Trash==1.8.3
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
stripe==13.0.1
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.33.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
yarg==0.1.9
